Redefined how GSA_RigidConstraints are treated, now a part of Structures_RigidLinks.
Rigid Constraints are now treated by BHoM as a Rigid Link object but with a fragment that if true tells GSA it is a Rigid Constraints. Removed previously added RigidConstraint Type and ENUM from this PR and instead adjusted the RigidLink methods. Also added more possible LinkConstraint options to the FromGsa - LinkConstraint method.